Completely change the API
The motivation is to make it easier to use commit outside the components or views. I rarely put event handler code in those functions themselves so it's better to have commit be something that is accessible earlier. Also trying to move toward referring to them as views instead of components.